Made the Auth Service Dynamic
I wanted to have the ability to have a debug auth service and a release that actually used a backend.  Right now I am using the environment production varialbe to determine this at run time.

I created the AuthInterface and the associated AuthDebugService class.  The Auth Service class reads the enironment setup and initializes the service variable accordingly.  Right now we only have debug, but soon we will have a web api back end.

Now I will have to do this for all the other services =)